podcast/requirements.txt
google-genai 
pydub